requirements.txt
langchain
langchain-core
langchain-community
langchain-openai
langchain_chroma
langchain_ollama
langgraph
chromadb
rank-bm25
sentence-transformers
huggingface-hub
numpy
python-dotenv
pypdf
openai
tavily-python
markdown
xhtml2pdf
reportlab